+++ app/models/concerns/replyable.rb
@@ -11,8 +11,12 @@ module Replyable
     has_many :reply_users
   end
 
-  def last_page
-    ((replies.where(reply_id: nil).count - 1) / Reply::REPLIES_PER_PAGE) + 1
+  def last_page(current_user)
+    if current_user&.admin?
+      ((replies.where(reply_id: nil).count - 1) / Reply::REPLIES_PER_PAGE) + 1
+    else
+      ((replies.where(reply_id: nil).shadow_banned(current_user).count - 1) / Reply::REPLIES_PER_PAGE) + 1
+    end
   end
 
   def get_or_create_subscription(current_user); end